refactored utils to src
