+++ src/print_5_executed_operations.py
